requirements.txt
# Core Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic>=2.6.1
pydantic-settings>=2.1.0

# Google AI (Gemini)
google-generativeai>=0.8.0

# Async Support
aiohttp>=3.9.3
# asyncio is built into Python 3, no package needed
httpx==0.26.0

# Data Validation & Processing
python-dateutil==2.8.2
pytz==2024.1

# Caching
redis==5.0.1
aioredis==2.0.1

# Logging & Monitoring
structlog==24.1.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-gcp-trace==1.6.0
opentelemetry-instrumentation-fastapi==0.43b0
prometheus-client==0.19.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.24.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx==0.26.0

# Development
python-dotenv==1.0.1
black==24.1.1
isort==5.13.2
mypy>=1.8.0
ruff==0.2.0

# Type stubs
types-redis==4.6.0.20240106
types-python-dateutil==2.8.19.20240106
